Added minimal support on firebase cloud messaging service --Refined UI feedback for chatActivity

+++ app/src/main/java/com/example/reconcile/ViewModel/data/message.kt
@@ -1,4 +1,4 @@
 package com.example.reconcile.ViewModel.data
 
-class message(val time : Int = 0, val message : String = "default", val ownerName : String = "no_name") {
+class message(val time : Long = 0, val message : String = "default", val ownerName : String = "no_name") {
 }
